# swagger.yaml
openapi: 3.0.3
info:
  title: Grymey
  description: >-
    # Grymey api documentation
  version: 1.0.0
  contact: { email: ammuftau74@gmail.com }
servers:
  - url: https://grymey.onrender.com/api/v1
    description: This is the original api URL
  - url: http://localhost:9000/api/v1
    description: This is my local Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /jars:
    get:
      tags:
        - 🔄 moneyjar
      summary: Get money jar
      description: >-
        The API endpoint retrieves all  money jar.
      operationId: getall money jar
      responses:
        '200':
          description: Get  money jar
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '1126'
            Date:
              schema:
                type: string
                example: Thu, 01 Jun 2023 05:44:27 GMT
            ETag:
              schema:
                type: string
                example: W/"466-tNv1SNtB6cuxJH2yZW7lrDnlA34"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                  message:
                    type: string
                    example: Money jar fetched successfully
                  success:
                    type: boolean
                    example: true    
    post:
      tags:
        - 🔄 moneyjar
      summary: create money jar
      description: create money jar.
      operationId: createmoneyjar
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: school fees
                description:
                  type: string
                  example: save some money for my school fees
                maturityDate:
                  type: string
                  example: '2025-09-30'
                targetAmount:
                  type: string
                  example: 300000
                autoDebit:
                  type: object
                  properties:
                    isActive:
                      type: boolean
                      example: true
                    amount:
                      type: number
                      example: 300
                    frequency:
                      type: string
                      enum: [daily, weekly, monthly, yearly]
                      example: daily
                type:
                  type: string
                  enum: [individual, group]
                  example: individual
      responses:
        '200':
          description: create money jar.
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '105'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:56:15 GMT
            ETag:
              schema:
                type: string
                example: W/"69-7OD21C6OqsuAipwphfFVCzbFDQc"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
                  message:
                    type: string
                    example: Money jar created successfully
                  success:
                    type: boolean
                    example: true
              examples:
                Forgot password request:
                  value:
                    data: {}
                    message: Money jar created successfully
                    success: true
...
  /jars/{jarId}/fund:
    post:
      tags:
        - 🔄 moneyjar
      summary: fund money jar
      description: fund money jar.
      operationId: fundmoneyjar
      parameters:
        - name: jarId
          in: path
          description: ID of the jar
          required: true 
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: string
                  example: 3000
      responses:
        '200':
          description: fund money jar.
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '105'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:56:15 GMT
            ETag:
              schema:
                type: string
                example: W/"69-7OD21C6OqsuAipwphfFVCzbFDQc"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
                  message:
                    type: string
                    example: Money jar funded successfully
                  success:
                    type: boolean
                    example: true
              examples:
                Forgot password request:
                  value:
                    data: {}
                    message: Money jar funded successfully
                    success: true
...
  /jars/{jarId}/lock:
    post:
      tags:
        - 🔄 moneyjar
      summary: lock money jar
      description: lock money jar.
      operationId: lockmoneyjar
      parameters:
        - name: jarId
          in: path
          description: ID of the jar
          required: true
      responses:
        '200':
          description: lock money jar.
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '105'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:56:15 GMT
            ETag:
              schema:
                type: string
                example: W/"69-7OD21C6OqsuAipwphfFVCzbFDQc"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
                  message:
                    type: string
                    example: Money jar lock successfully
                  success:
                    type: boolean
                    example: true
              examples:
                Forgot password request:
                  value:
                    data: {}
                    message: Money jar lock successfully
                    success: true
  /jars/{jarId}/unlock:
    post:
      tags:
        - 🔄 moneyjar
      summary: unlock money jar
      description: unlock money jar.
      operationId: unlockmoneyjar
      parameters:
        - name: jarId
          in: path
          description: ID of the jar
          required: true
      responses:
        '200':
          description: unlock money jar.
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '105'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:56:15 GMT
            ETag:
              schema:
                type: string
                example: W/"69-7OD21C6OqsuAipwphfFVCzbFDQc"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
                  message:
                    type: string
                    example: Money jar unlocked successfully
                  success:
                    type: boolean
                    example: true
              examples:
                Forgot password request:
                  value:
                    data: {}
                    message: Money jar unlocked successfully
                    success: true
...